+++ src/main/java/com/themetalstorm/bibliothekssystem/repository/GenreRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface GenreRepository extends JpaRepository<Genre, Long> {
+public interface GenreRepository extends JpaRepository<Genre, Integer> {
     boolean existsByName(String name);
 
     Optional<Genre> findByName(String name);
